python/src/requirements.txt
flask
python-logstash